Moved KapuaErrorJson to kapua-service-commons
